Move Datatore and WindowData to separate files.
